drivers: gpu-kgsl: Increase worker thread priority
Currently, the kgsl worker thread is erroneosly ranked right below
Android's audio threads in terms of priority.

The kgsl worker thread is in the critical path for rendering frames to the
display, so increase its priority to match the priority of MDSS' kthread (mdss_fb0).

Aquarius223: Reduce the default 16 to 6, fix the screen tearing
             in some scenes.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I0f47d6cc4a0b128f3b2246f2d80c71e0ae7c27c4